+++ src/onegov.town/onegov/town/models/__init__.py
@@ -1,4 +1,5 @@
 from onegov.town.models.editor import Editor
+from onegov.town.models.file import File, FileCollection
 from onegov.town.models.form import BuiltinFormDefinition, CustomFormDefinition
 from onegov.town.models.image import Image, ImageCollection, Thumbnail
 from onegov.town.models.page import News, Topic
@@ -9,6 +10,8 @@ __all__ = [
     'BuiltinFormDefinition',
     'CustomFormDefinition',
     'Editor',
+    'File',
+    'FileCollection',
     'Image',
     'ImageCollection',
     'News',